fastAPI/requirements.txt
fastapi==0.115.8
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.8.3
gtts==2.4.0
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
torch
transformers
IndicTransToolkit
sentencepiece
jwt
langchain_groq
langchain
langchain_community
